requirements.txt
langchain
langchain-google-genai
langchain-mongodb
pymongo
flask
pypdf
python-dotenv